[RFC] Update spec support level description (#6842)
* update the Config type to change support level representation

this also starts the process of removing builds. no spec supports them,
so this is just removing a bunch of uses of the parameter. i am not
removing them from the URLs yet because that would cause a bunch of URL
breakage

* add unmaintained variant

* update config files

* further fixes

* maybe (?) preemptively fix e2e

its a pain to run on this computer so...i'm not

* consistency

* update about test and footer text for foundation support

* add tooltip explaining "Foundational Support" to footer