refactor APL code to allow use within an Analyzer.
some features are not supported in this context. notably: lookaheads.

this will error if an APL tries to use a lookahead in an Analyzer context.

looking through usage, it appears that only Devastation and Protection Paladin
use lookaheads still. Paladin is disabled and Devastation's APL is disabled,
so the feature could probably be removed.

consider this one step on the road to full Analyzer-ization