use an APL to estimate unavoidable downtime for Cata Arms
Arms is a downtime spec in Cata: you do not have enough cooldowns
or Rage to fill every GCD, so some are unavoidably empty.

this uses an APL-based approach to attempt to identify situations
where there isn't an appropriate spell to cast. such periods are then
added to the `AlwaysBeCasting` as "uptime" regions, boosting overall uptime