update Brewmaster APL and simplify Face Palm tracking
this is more deletion than addition.

I cut the PtA APL because the talent is basically dead. Playing it optimally is complicated plate-spinning, and playing it sub-optimally is not what this should be recommending.

I simplified the FP proc detection to just assume the average case. The existing method was quietly broken and had been for a while since apparently AP is no longer logged on damage events (only casts). But there are even worse issues with it (like: boss damage amps break it, debuff damage amps break it, etc). Assuming the average helps avoid CD errors, but since the CDR amount is so small it isn't a big factor in overall stats.

I also adjusted which problems are shown by default in the APL section to prefer examples that come *after* the first 10s of the fight. The reasoning here is that openers often behave differently and may not be properly accounted for in the APL. Brewmaster in particular has a "throw everything at the boss" opener with no real rules where you just aim to get everything on cooldown. If the only examples of a mistake are early in a fight, those will still be shown.