adjust spellgen to avoid `readonly` problem
tl;dr - adding `satisfies RetailSpell` fixes it.

in order for that to work, some internal fields needed
to be stripped from the data. they aren't part of the public interface
from wow-dbc so wow-dbc should *probably* be stripping them but
it is tricky to do it generically. still, probably better to have wow-dbc
do it statically than wowa