fix e2e failures
these are primarily noise from the midnight changes to the header,
along with changes to an error with dom nesting that we were ignoring before.

i've fixed the remaining 2 issues with `validateDOMNesting`, and removed
the ignore

for the moment, I also enabled `handleExpansionChecker` so that previous-expansion
logs in configs pass e2e. this will get turned back off closer to midnight release